chore(refactoring): change same name function
- instance, bootstrap
